jotain poikkeuskäsittelyä

+++ src/main/java/rest/todo/repositories/TodoRepository.java
@@ -9,6 +9,7 @@ import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
 import rest.todo.entities.TodoEntity;
+import rest.todo.exceptions.NoContentException;
 
 @Transactional
 @Repository
@@ -36,15 +37,22 @@ public class TodoRepository implements ITodoRepository{
 	}
 
 	@Override
-	public TodoEntity updateTodoEntity(TodoEntity todoEntity) {
+	public TodoEntity updateTodoEntity(TodoEntity todoEntity) throws NoContentException {
 		TodoEntity tde = getTodoById(todoEntity.getId());
+		if(tde == null) {
+			throw new NoContentException ("Päivitettävää tietoa ei löytynyt");
+		}
 		tde.setTodo(todoEntity.getTodo());
 		 entityManager.flush();
 		 return tde;
 	}
 
 	@Override
-	public void deleteTodoEntity(Long todoId) {
+	public void deleteTodoEntity(Long todoId) throws NoContentException {
+		TodoEntity tde = getTodoById(todoId);
+		if(tde == null) {
+			throw new NoContentException ("Poistettavaa tietoa ei löytynyt");
+		}
 		entityManager.remove(getTodoById(todoId));
 	}
